feat: migrate usePermissions useState to jotai atoms
Moves permissions state and platform detection out of local useState
into shared jotai atoms in atoms/permissions.ts. Relocates
PermissionState and PermissionStatus types to the atoms file and re-
exports them from the hook for backward compat.

isMacOSAtom (atoms/app.ts) is now shared between usePermissions and
useScreenRecorder, so platform detection happens once and both hooks
read the same value.
